Rename types to @types
