+++ src/@types/switches.ts
