+++ src/@types/window.d.ts
